backend/requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
cryptography==37.0.2
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==3.2.7
django-allauth==0.50.0
django-bootstrap-v5==1.0.11
django-cors-headers==3.12.0
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
executing==0.8.3
idna==3.3
ipython==8.2.0
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
numpy==1.22.3
oauthlib==3.2.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.29
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
tomli==1.2.3
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.9
wcwidth==0.2.5